pyproject.toml
[tool.poetry]
name = "tg-react"
version = "3.1.0"
description = "Helpers for react based applications running on django."
authors = [
    "Thorgate <code@thorgate.eu>",
    "Jyrno Ader <jyrno@thorgate.eu>",
    "Jorgen Ader <jorgen@thorgate.eu>",
]
license = "BSD"
homepage = "https://github.com/thorgate/tg-react"
repository = "https://github.com/thorgate/tg-react"
readme = "README.md"
packages = [
    { include = "tg_react" },
]
include = [
    "tg_react/**/*.txt",
    "tg_react/**/*.html",
    "tg_react/**/*.mo",
    "tg_react/**/*.po"
]
keywords = ["tg-react", "thorgate", "rest", "drf"]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Framework :: Django',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: BSD License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
]

[tool.poetry.dependencies]
django = ">=2.2, <4.2"
djangorestframework = ">=3.9.2"

python = ">=3.7,<4"

[tool.poetry.dev-dependencies]
pytest = "==7.*"
coverage = "*"
coveralls = "*"
pytest-cov = "*"
pytest-django = "*"
pytest-xdist = "*"
model-bakery = "*"
black = "==22.3.0"
prospector = "*"
sphinx = "==3.*"
tox = "*"
tox-poetry = "*"
tox-gh-actions = "*"

example = { path = "./demo", develop = true }

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 140
require-code = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "dummy_settings"
django_find_project = "false"
python_files = "tests/*.py tests/**/*.py"
norecursedirs = "venv* dist* htmlcov* example* .tox*"
addopts = "--capture=no -n auto"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''

(
  \.eggs         # exclude a few common directories in the
| \.git          # root of the project
| \.mypy_cache
| \.tox
| \.venv
| _build
| build
| dist
| docs
)
'''
